Move nodes to own folder
