+++ client/src/context/GraphProvider.tsx
@@ -1,0 +1,14 @@
+// SPDX-FileCopyrightText: 2024 The Aalto Grades Developers
+//
+// SPDX-License-Identifier: MIT
+
+import {Dispatch, SetStateAction, createContext} from 'react';
+
+export type NodeValues = {[key: string]: number};
+type NodeValuesContext = {
+  nodeValues: NodeValues;
+  setNodeValues: Dispatch<SetStateAction<NodeValues>>;
+};
+export const NodeValuesContext = createContext<NodeValuesContext>(
+  {} as NodeValuesContext
+);
